requirements.txt
uvicorn
fastapi
garminconnect
python-dotenv